requirements.txt
ipykernel
litellm
transformers
openai
tqdm
python-dotenv
